refactor(purchase-order): HqPurchaseOrderView 책임 단위 컴포넌트 분리
본사 발주 목록 화면 809 → 404 (-50%, -405). root CLAUDE.md 의 SFC ~500
가이드 도달. 인라인 SVG 아이콘·상태 라벨 매핑·발주 상세 패널·취소 모달
4개 영역을 책임 단위로 떼어내 부모는 layout + 상태 조립만.

- icons.js 에 XIcon/EditIcon/InfoIcon/WarehouseIcon/UserIcon/ZapIcon/
  TruckIcon 7개 추가 (PR #416 의 기존 모듈에 통합).
- useStatusFormat composable: statusClass/statusLabel/historyDotClass/
  historyTextClass/formatDate 재사용 가능 형태로. 본사 시점 라벨 7단계.
- PurchaseOrderDetailPanel: 우측 발주 상세 패널 (헤더/기본정보/품목 테이블/
  진행 이력/액션 버튼/하단 닫기). order props + close/edit/cancel emit.
  REQUESTED 단계만 본사 권한(수정/취소) 노출 — 단계 분기 자식이 책임.
- PurchaseOrderCancelModal: cancelReason textarea 모달. open/order props +
  cancel/confirm emit. reason 은 모달 로컬 state.
- 부모 view 의 인라인 IconBase 정의 + 5개 헬퍼 + 235줄 상세 패널 +
  55줄 취소 모달 모두 제거. 사용 안 하는 import 5개(EditIcon/InfoIcon/
  UserIcon/WarehouseIcon/XIcon) 도 정리.
