+++ src/components/hq/purchase-order/icons.js
@@ -57,3 +57,41 @@ export const AlertTriangleIcon = IconBase([
   { tag: 'path', attrs: { d: 'M12 9v4' } },
   { tag: 'path', attrs: { d: 'M12 17h.01' } },
 ])
+
+export const XIcon = IconBase([
+  { tag: 'path', attrs: { d: 'M18 6 6 18' } },
+  { tag: 'path', attrs: { d: 'm6 6 12 12' } },
+])
+
+export const EditIcon = IconBase([
+  { tag: 'path', attrs: { d: 'M11 4H4a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2v-7' } },
+  { tag: 'path', attrs: { d: 'M18.5 2.5a2.121 2.121 0 0 1 3 3L12 15l-4 1 1-4 9.5-9.5Z' } },
+])
+
+export const InfoIcon = IconBase([
+  { tag: 'circle', attrs: { cx: '12', cy: '12', r: '9' } },
+  { tag: 'path', attrs: { d: 'M12 10v6' } },
+  { tag: 'path', attrs: { d: 'M12 7h.01' } },
+])
+
+export const WarehouseIcon = IconBase([
+  { tag: 'path', attrs: { d: 'M3 10.5 12 4l9 6.5' } },
+  { tag: 'path', attrs: { d: 'M5 9.5V20h14V9.5' } },
+  { tag: 'path', attrs: { d: 'M10 20v-5h4v5' } },
+])
+
+export const UserIcon = IconBase([
+  { tag: 'path', attrs: { d: 'M20 21a8 8 0 0 0-16 0' } },
+  { tag: 'circle', attrs: { cx: '12', cy: '8', r: '4' } },
+])
+
+export const ZapIcon = IconBase([
+  { tag: 'polygon', attrs: { points: '13 2 3 14 12 14 11 22 21 10 12 10 13 2' } },
+])
+
+export const TruckIcon = IconBase([
+  { tag: 'path', attrs: { d: 'M10 17H5a2 2 0 0 1-2-2V7h11v10Z' } },
+  { tag: 'path', attrs: { d: 'M14 17h-1V9h3l3 3v5h-1' } },
+  { tag: 'circle', attrs: { cx: '7.5', cy: '17.5', r: '1.5' } },
+  { tag: 'circle', attrs: { cx: '17.5', cy: '17.5', r: '1.5' } },
+])
